fix(framework): complete the git-dep walk, extract dep_walk module
The cherry-picked walk missed two pieces on the rebased branch: dev
and build deps leaked into the metadata results, and extension
discovery never saw git deps because it had moved to the new
direct-only walk. Filter the resolve tree to normal-kind edges, merge
depth-1 metadata deps into discovery, and run cargo metadata with
--offline so macro expansion never touches the network.

Move the walk machinery into a dep_walk module with its tests. Share
the duplicate-name source label across the three producers. Update
the CLI walk tests to the fallback semantics.
