+++ spel-cli/src/generate_idl.rs
@@ -371,7 +371,11 @@ mod tests {
     }
 
     #[test]
-    fn find_path_dep_dirs_ignores_registry_and_git_deps() {
+    fn find_path_dep_dirs_falls_back_to_path_only_when_metadata_fails() {
+        // The unfetchable git dep makes `cargo metadata --offline` fail, so
+        // the walk warns and degrades to path-only results. Registry and git
+        // deps are then absent not because they are ignored, but because the
+        // metadata source for them is unavailable.
         let tmp = TempDir::new("find-path-deps-filter");
 
         tmp.write(
@@ -392,11 +396,11 @@ mod tests {
 
         let result = find_path_dep_dirs(&program);
         assert!(
-            result.warnings.is_empty(),
-            "unexpected warnings: {:?}",
+            result.warnings.iter().any(|w| w.contains("cargo metadata")),
+            "expected a `cargo metadata` warning, got: {:?}",
             result.warnings
         );
-        // Only the path dep (core) should be returned, not serde or nssa_core
+        // The path dep (core) is still returned via the manifest walk.
         assert_eq!(result.dirs.len(), 1);
         assert!(result.dirs[0].ends_with("core"));
     }
@@ -603,6 +607,7 @@ mod tests {
             "[package]\nname = \"myprog\"\nversion = \"0.1.0\"\nedition = \"2021\"\n\n\
              [dependencies]\ncommon = { path = \"../../libs/common\" }\n",
         );
+        tmp.write("programs/myprog/src/lib.rs", "");
         // Source file is in a deeply nested dir with no intermediate Cargo.toml.
         let program = tmp.write("programs/myprog/src/deep/nested/token.rs", "");
