+++ spel-framework-core/src/idl_gen.rs
@@ -17,7 +17,11 @@ use crate::idl::{IdlAccountItem, IdlArg, IdlInstruction, IdlPda, IdlSeed, SpelId
 
 use crate::account_types::{collect_account_types, syn_type_to_idl_type};
 
-use crate::extension::{check_duplicate_instruction_names, discover_extension_instructions};
+use crate::extension::{
+    check_duplicate_instruction_names, discover_extension_instructions, instruction_source_label,
+};
+
+pub use crate::dep_walk::find_path_dep_dirs;
 
 /// Error type returned by [`generate_idl_from_file`].
 #[derive(Debug)]
@@ -165,13 +169,7 @@ fn generate_idl_inner(
         check_duplicate_instruction_names(instructions.iter().map(|i| {
             (
                 i.fn_name.to_string(),
-                match &i.external_call_path {
-                    Some(p) => match p.segments.first() {
-                        Some(seg) => format!("extension {}", seg.ident),
-                        None => "an extension".to_string(),
-                    },
-                    None => "this module".to_string(),
-                },
+                instruction_source_label(i.external_call_path.as_ref()),
             )
         }))
         .map_err(IdlGenError::DuplicateInstruction)?;
@@ -820,383 +818,6 @@ fn parse_single_pda_seed(call: &syn::ExprCall) -> Result<PdaSeedDef, syn::Error>
     }
 }
 
-// ─── Dependency scanning (shared by CLI and proc-macro) ─────────────
-
-/// Return the crate-root directories of every runtime dependency of the
-/// `Cargo.toml` nearest to `source_path`.
-///
-/// Path dependencies are resolved transitively via a direct manifest walk
-/// (fast, no subprocess). Git and registry dependencies are additionally
-/// enumerated via a single `cargo metadata --format-version 1` call so that
-/// extensions shipped through git URLs or crates.io remain discoverable by
-/// name-based lookups (extension attrs, wrap specs, inject specs). Results
-/// from both sources are merged and deduplicated by canonical path.
-///
-/// **Transitive path-dependencies** are resolved: if a discovered path
-/// dependency itself declares further path-based dependencies, those are
-/// included as well (with cycle detection).
-///
-/// `[dev-dependencies]` and `[build-dependencies]` are deliberately
-/// excluded: types defined in those crates are not part of the program's
-/// on-chain interface and must not appear in the generated IDL. The
-/// `cargo metadata` merge only reads the runtime resolve tree, so dev and
-/// build deps are naturally filtered out.
-///
-/// In workspace projects the function detects when the nearest `Cargo.toml`
-/// is a workspace root manifest and searches for the actual crate manifest
-/// containing `[dependencies]`; `cargo metadata` is invoked against that
-/// member manifest so its dep graph is what gets enumerated.
-///
-/// `on_warning` is called for non-fatal issues (missing dep directories,
-/// unparseable manifests, `cargo metadata` failures, etc.). Pass `|_| {}`
-/// to ignore warnings. If `cargo metadata` fails or is unavailable the
-/// function silently returns only the path-dep results so downstream
-/// expansion stays deterministic.
-pub fn find_path_dep_dirs<F: FnMut(String)>(source_path: &Path, mut on_warning: F) -> Vec<PathBuf> {
-    let manifest = match _find_crate_manifest(source_path, &mut on_warning) {
-        Some(m) => m,
-        None => return vec![],
-    };
-
-    let content = match std::fs::read_to_string(&manifest) {
-        Ok(c) => c,
-        Err(e) => {
-            on_warning(format!(
-                "⚠️  could not read manifest '{}': {}",
-                manifest.display(),
-                e
-            ));
-            return vec![];
-        },
-    };
-    let value: toml::Value = match toml::from_str(&content) {
-        Ok(v) => v,
-        Err(e) => {
-            on_warning(format!(
-                "⚠️  failed to parse manifest '{}': {}",
-                manifest.display(),
-                e
-            ));
-            return vec![];
-        },
-    };
-
-    let manifest_dir = match manifest.parent() {
-        Some(d) => d.to_path_buf(),
-        None => return vec![],
-    };
-
-    // Check if this is a workspace root — if so, it has no [dependencies] of its
-    // own.  We need to find the actual crate manifest for the program binary.
-    let is_workspace = value.get("workspace").is_some() && value.get("package").is_none();
-
-    let mut dirs = Vec::new();
-    let mut visited = HashSet::new();
-    let metadata_manifest: Option<PathBuf> = if is_workspace {
-        let member = _find_member_manifest(&manifest_dir, &value, source_path, &mut on_warning);
-        if let Some(m) = &member {
-            _resolve_path_deps_recursive(&m, &mut dirs, &mut visited, &mut on_warning);
-        }
-        member
-    } else {
-        _resolve_path_deps_recursive(&manifest, &mut dirs, &mut visited, &mut on_warning);
-        Some(manifest.clone())
-    };
-
-    // Merge in git and registry dep dirs via `cargo metadata`. Best-effort:
-    // any failure falls back to path-only results.
-
-    if let Some(m) = &metadata_manifest {
-        for dir in _find_dep_dirs_via_cargo_metadata(m, &mut on_warning) {
-            let canonical = dir.canonicalize().unwrap_or_else(|_| dir.clone());
-            if visited.insert(canonical) {
-                dirs.push(dir);
-            }
-        }
-    }
-
-    dirs
-}
-
-/// Recursively extract path-based dependencies from a manifest, following
-/// transitive path deps.  `visited` tracks canonicalised directories to avoid
-/// infinite loops.
-fn _resolve_path_deps_recursive<F: FnMut(String)>(
-    manifest: &Path,
-    dirs: &mut Vec<PathBuf>,
-    visited: &mut HashSet<PathBuf>,
-    on_warning: &mut F,
-) {
-    let manifest_dir = match manifest.parent() {
-        Some(d) => d.to_path_buf(),
-        None => return,
-    };
-
-    // Deduplicate by canonical path.
-    let canonical = match &manifest_dir.canonicalize() {
-        Ok(c) => c.clone(),
-        Err(_) => manifest_dir.clone(),
-    };
-    if !visited.insert(canonical) {
-        return; // already processed — cycle or duplicate
-    }
-
-    let content = match std::fs::read_to_string(manifest) {
-        Ok(c) => c,
-        Err(e) => {
-            on_warning(format!(
-                "⚠️  could not read manifest '{}': {}",
-                manifest.display(),
-                e
-            ));
-            return;
-        },
-    };
-    let value: toml::Value = match toml::from_str(&content) {
-        Ok(v) => v,
-        Err(e) => {
-            on_warning(format!(
-                "⚠️  failed to parse manifest '{}': {}",
-                manifest.display(),
-                e
-            ));
-            return;
-        },
-    };
-
-    // Skip workspace roots — they have no [dependencies].
-    if value.get("workspace").is_some() && value.get("package").is_none() {
-        return;
-    }
-
-    if let Some(table) = value.get("dependencies").and_then(|v| v.as_table()) {
-        for (name, dep) in table {
-            if let Some(rel) = dep.get("path").and_then(|v| v.as_str()) {
-                let dep_dir = manifest_dir.join(rel);
-                if !dep_dir.is_dir() {
-                    on_warning(format!(
-                        "⚠️  path dependency '{}' points to non-existent directory: {}",
-                        name,
-                        dep_dir.display()
-                    ));
-                    continue;
-                }
-                // Deduplicate by canonical path.
-                let canonical = match &dep_dir.canonicalize() {
-                    Ok(c) => c.clone(),
-                    Err(_) => dep_dir.clone(),
-                };
-                if visited.contains(&canonical) {
-                    continue;
-                }
-                dirs.push(dep_dir.clone());
-
-                // Recurse into the dependency's own Cargo.toml for transitive deps.
-                let dep_manifest = dep_dir.join("Cargo.toml");
-                if dep_manifest.exists() {
-                    _resolve_path_deps_recursive(&dep_manifest, dirs, visited, on_warning);
-                }
-            }
-        }
-    }
-}
-
-/// Given a workspace root directory, try to locate the member crate manifest
-/// that contains `source_path`.
-fn _find_member_manifest<F: FnMut(String)>(
-    workspace_root: &Path,
-    workspace_value: &toml::Value,
-    source_path: &Path,
-    on_warning: &mut F,
-) -> Option<PathBuf> {
-    // Try to get the explicit member list from [workspace.members].
-    let members: Vec<String> = workspace_value
-        .get("workspace")
-        .and_then(|w| w.get("members"))
-        .and_then(|m| m.as_array())
-        .map(|arr| {
-            arr.iter()
-                .filter_map(|v| v.as_str())
-                .map(|s| s.to_string())
-                .collect()
-        })
-        .unwrap_or_default();
-
-    // Expand glob patterns (e.g. "crates/*") into concrete directories.
-    let concrete_members: Vec<String> = if members.iter().any(|m| m.contains('*')) {
-        let mut expanded = Vec::new();
-        for pattern in &members {
-            if pattern.contains('*') {
-                // Simple glob expansion: replace * with readdir.
-                let prefix = pattern.split_once('*').map(|(p, _)| p).unwrap_or("");
-                let dir = workspace_root.join(prefix);
-                if let Ok(entries) = std::fs::read_dir(&dir) {
-                    for entry in entries.flatten() {
-                        if entry.file_type().map_or(true, |ft| ft.is_dir()) {
-                            expanded.push(format!(
-                                "{}/{}",
-                                prefix,
-                                entry.file_name().to_string_lossy()
-                            ));
-                        }
-                    }
-                }
-            } else {
-                expanded.push(pattern.clone());
-            }
-        }
-        expanded
-    } else {
-        members.clone()
-    };
-
-    // Find the member whose directory contains source_path.
-    let source_dir = source_path.parent().unwrap_or(source_path);
-    for member in &concrete_members {
-        let member_dir = workspace_root.join(member.as_str());
-        if member_dir.is_dir() && source_dir.starts_with(&member_dir) {
-            let manifest = member_dir.join("Cargo.toml");
-            if manifest.exists() {
-                return Some(manifest);
-            }
-        }
-    }
-
-    // Fallback: recursively search all subdirectories for a Cargo.toml that
-    // contains source_path.  This handles nested workspace members (e.g.
-    // `methods/guest`) when the explicit `members` list is absent/mismatched.
-    on_warning(format!(
-        "⚠️  workspace at '{}' has no matching member for '{}'; searching all subdirectories",
-        workspace_root.display(),
-        source_path.display()
-    ));
-
-    fn _search_recursive(dir: &Path, target_dir: &Path) -> Option<PathBuf> {
-        // Search children FIRST (depth-first), then check current dir.
-        // This ensures we find the deepest matching member manifest rather
-        // than returning the workspace root immediately.
-        if let Ok(entries) = std::fs::read_dir(dir) {
-            for entry in entries.flatten() {
-                if entry.file_type().is_ok_and(|ft| ft.is_dir()) {
-                    if let Some(found) = _search_recursive(&entry.path(), target_dir) {
-                        return Some(found);
-                    }
-                }
-            }
-        }
-        // Check current dir — but skip virtual workspace manifests (no [package]).
-        let manifest = dir.join("Cargo.toml");
-        if manifest.exists() && target_dir.starts_with(dir) {
-            // Skip virtual workspace manifests that have [workspace] but no [package].
-            let is_virtual_workspace = std::fs::read_to_string(&manifest)
-                .ok()
-                .and_then(|content| content.parse::<toml::Value>().ok())
-                .map(|v| v.get("workspace").is_some() && v.get("package").is_none())
-                .unwrap_or(false);
-            if !is_virtual_workspace {
-                return Some(manifest);
-            }
-        }
-        None
-    }
-
-    _search_recursive(workspace_root, source_dir)
-}
-
-/// Walk up from `start` to find the nearest `Cargo.toml`.
-pub(crate) fn _find_crate_manifest<F: FnMut(String)>(
-    start: &Path,
-    on_warning: &mut F,
-) -> Option<PathBuf> {
-    let mut dir: &Path = if start.is_file() {
-        start.parent()?
-    } else {
-        start
-    };
-    loop {
-        let candidate = dir.join("Cargo.toml");
-        if candidate.exists() {
-            return Some(candidate);
-        }
-        dir = match dir.parent() {
-            Some(p) => p,
-            None => {
-                on_warning(format!(
-                    "⚠️  no Cargo.toml found walking up from '{}'",
-                    start.display()
-                ));
-                return None;
-            },
-        };
-    }
-}
-
-fn _find_dep_dirs_via_cargo_metadata<F: FnMut(String)>(
-    manifest:&Path,
-    on_warning: &mut F,
-) -> Vec<PathBuf> {
-    let output = match std::process::Command::new("cargo")
-        .args(["metadata", "--format-version", "1", "--manifest-path"])
-        .arg(manifest)
-        .output()
-    {
-        Ok(o) => o,
-        Err(e) => {
-            on_warning(format!("could not run `cargo metadata`: {e}"));
-            return Vec::new();
-        }
-    };
-    if !output.status.success() {
-        on_warning(format!("`cargo metadata` failed: {}",
-            String::from_utf8_lossy(&output.stderr).trim()
-        ));
-        return Vec::new();
-    }
-    
-    let meta: serde_json::Value = match serde_json::from_slice(&output.stdout) {
-        Ok(v) => v,
-        Err(e) => {
-            on_warning(format!("failed to parse metadata: {e}"));
-            return Vec::new();
-        }
-    };
-
-    let workspace_members: HashSet<String> = meta
-        .get("workspace_members")
-        .and_then(|v| v.as_array())
-        .map(|arr| arr.iter().filter_map(|v| v.as_str().map(String::from)).collect())
-        .unwrap_or_default();
-
-    let source_canonical = manifest.canonicalize().ok();
-    let mut dirs = Vec::new();
-
-    let empty = Vec::new();
-    let packages = meta.get("packages").and_then(|v| v.as_array()).unwrap_or(&empty);
-    for pkg in packages {
-        if let Some(id) = pkg.get("id").and_then(|v| v.as_str()) {
-            if workspace_members.contains(id) {
-                continue;
-            }
-        }
-
-        let Some(mp) = pkg.get("manifest_path").and_then(|v| v.as_str()) else {
-            continue;
-        };
-        let pkg_manifest = PathBuf::from(mp);
-
-        if let (Some(src), Ok(pkg_c)) = (&source_canonical, pkg_manifest.canonicalize()) {
-            if src == &pkg_c {
-                continue;
-            }
-        }
-        if let Some(dir) = pkg_manifest.parent() {
-            dirs.push(dir.to_path_buf());
-        }
-    }
-
-    dirs
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -1809,88 +1430,5 @@ mod tests {
         assert!(matches!(&fields[5].type_, IdlType::Primitive(s) if s == "account_id"));
         assert!(matches!(&fields[6].type_, IdlType::Option { .. }));
         assert!(matches!(&fields[7].type_, IdlType::Vec { .. }));
-    }
-
-    #[test]
-    fn find_path_dep_dirs_dedups_diamond_graph() {
-        // Diamond dep graph: sample → ext_a (direct), sample → ext_b → ext_a (transitive).
-        // The buggy version of `_resolve_path_deps_recursive` pushes ext_a to the
-        // returned Vec twice — once via the direct edge, once via ext_b's transitive
-        // edge — because the push happens before the visited-set check in the
-        // recursive call. Latent on `main` because downstream callers
-        // (`collect_items_from_crate_dirs`) have a second dedup layer keyed by
-        // canonical source-file path, but a caller that skips that layer sees the
-        // duplicate dir and produces duplicate items.
-        //
-        // This test reproduces the diamond and asserts the returned Vec contains
-        // no duplicates by canonical path. Fails on buggy code, passes on the fix.
-
-        let tmp = TempDir::new("diamond");
-
-        tmp.write(
-            "ext-a/Cargo.toml",
-            r#"
-[package]
-name = "ext-a"
-version = "0.1.0"
-edition = "2021"
-"#,
-        );
-        tmp.write("ext-a/src/lib.rs", "");
-
-        tmp.write(
-            "ext-b/Cargo.toml",
-            r#"
-[package]
-name = "ext-b"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-ext-a = { path = "../ext-a" }
-"#,
-        );
-        tmp.write("ext-b/src/lib.rs", "");
-
-        tmp.write(
-            "sample/Cargo.toml",
-            r#"
-[package]
-name = "sample"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-ext-a = { path = "../ext-a" }
-ext-b = { path = "../ext-b" }
-"#,
-        );
-        tmp.write("sample/src/lib.rs", "");
-
-        let dirs = find_path_dep_dirs(&tmp.path().join("sample"), |_| {});
-
-        // Canonicalise every returned dir, count unique canonical paths.
-        // The dirs Vec MUST have no duplicate canonical paths — including ext-a,
-        // which is reachable via both the direct edge and ext-b's transitive edge.
-        let unique: std::collections::HashSet<PathBuf> = dirs
-            .iter()
-            .map(|p| p.canonicalize().unwrap_or_else(|_| p.clone()))
-            .collect();
-
-        assert_eq!(
-            unique.len(),
-            dirs.len(),
-            "find_path_dep_dirs returned duplicate canonical paths: {:?}",
-            dirs
-        );
-
-        // Also assert ext-a is in the result (paranoia: the diamond graph
-        // actually got walked, not just an empty result).
-        let ext_a_canonical = tmp.path().join("ext-a").canonicalize().unwrap();
-        assert!(
-            unique.contains(&ext_a_canonical),
-            "ext-a not in result; diamond walk failed: {:?}",
-            dirs
-        );
     }
 }
